+++ R/AddScaleBar.R
@@ -3,95 +3,142 @@
 #'
 #' This function can be used to add a scale bar to a plot.
 #'
-#' @param asp 'numeric'.
-#'   \emph{y/x} aspect ratio for spatial axes.
 #' @param unit 'character'.
-#'   Axis unit of measurement, for example "METERS".
-#' @param is.lonlat 'logical'.
-#'   If true, plot coordinates are in longitude and latitude.
+#'   Label describing the unit of measurement of scale distances, for example "METERS".
+#' @param conv.fact 'numeric'.
+#'   Conversion factor for changing the unit of measurement for scale distances.
+#'   For example, if user coordinates of the plotting region are in meters,
+#'   specify \code{3.28084} to display scale distances in feet.
+#' @param vert.exag 'logical', 'numeric', or 'character'.
+#'   Either a logical value indicating whether to include a vertical exaggeration label;
+#'   or a custom \emph{y/x} aspect ratio to include in this label.
+#' @param longlat 'logical'.
+#'   Whether user coordinates of the plotting region are in longitude and latitude;
+#'   if true, scale distances are in kilometers.
+#'   Note that scale distances are calculated at the maps latitude midpoint
+#'   using the Great Circle distance (WGS84 ellipsoid) method.
 #' @param loc 'character'.
-#'   Position of the scale bar in the plot region:
-#'   "bottomleft", "topleft", "topright", or "bottomright" to denote scale location.
+#'   Position of the scale bar in the plot region specified by keyword:
+#'   \code{"bottomleft"}, \code{"topleft"}, \code{"topright"}, or \code{"bottomright"}.
 #' @param offset 'numeric'.
-#'   The x and y adjustments of the scale bar, in inches.
-#' @param lab.vert.exag 'logical'.
-#'   If true, a label is drawn specifying the vertical exaggeration.
+#'   Vector of length 2 indicating the \emph{x} and \emph{y} adjustments
+#'   to the position of the scale bar, in inches.
 #'
 #' @return Used for the side-effect of a scale bar drawn on the current graphics device.
 #'
 #' @author J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
 #'
-#' @seealso \code{\link{PlotCrossSection}}, \code{\link{PlotMap}}
+#' @seealso \code{\link{PlotMap}}, \code{\link{PlotCrossSection}}
 #'
 #' @keywords hplot
 #'
 #' @export
 #'
 #' @examples
-#' plot(-100:100, -100:100, type = "n", xlab = "x", ylab = "y", asp = 2)
-#' AddScaleBar(2, unit = "FEET", loc = "topleft")
-#' AddScaleBar(2, unit = "METERS", loc = "bottomright", offset = c(-0.2, 0))
+#' plot(-100:100, -100:100, type = "n", xlab = "x in meters", ylab = "y in meters", asp = 2)
+#' AddScaleBar()
+#' AddScaleBar(unit = "METERS", loc = "topleft")
+#' AddScaleBar(unit = "METERS", vert.exag = TRUE, loc = "topright", offset = c(-0.2, 0))
+#' AddScaleBar(unit = "FEET", conv.fact = 3.28084, loc = "bottomright")
+#'
+#' plot(c(-38.31, -35.5), c(40.96, 37.5), type = "n", xlab = "longitude", ylab = "latitude")
+#' AddScaleBar(unit = "km", longlat = TRUE)
+#' AddScaleBar(unit = "mi", conv.fact = 0.621371, longlat = TRUE, loc = "topright")
 #'
 
-AddScaleBar <- function(asp=1, unit=NULL, is.lonlat=FALSE,
+AddScaleBar <- function(unit=NULL, conv.fact=NULL, vert.exag=NULL, longlat=FALSE,
                         loc=c("bottomleft", "topleft", "topright", "bottomright"),
-                        offset=c(0, 0), lab.vert.exag=NULL) {
+                        offset=NULL) {
+
+  checkmate::assertCharacter(unit, all.missing=FALSE, min.len=1, max.len=2, null.ok=TRUE)
+  checkmate::assertNumeric(conv.fact, all.missing=FALSE, min.len=1, max.len=2, null.ok=TRUE)
+  checkmate::qassert(vert.exag, c("B1", "S1", "N1", "0"))
+  checkmate::assertFlag(longlat)
   loc <- match.arg(loc)
-  usr <- graphics::par("usr")
-  pin <- graphics::par("pin")
+  checkmate::assertNumeric(offset, finite=TRUE, min.len=1, max.len=2, null.ok=TRUE)
 
-  if (is.null(asp)) asp <- pin[2] / pin[1]
+  pin  <- graphics::par("pin")   # plot dimensions in inches (width, height)
+  xaxp <- graphics::par("xaxp")  # extreme tick marks and number of intervals (x1, x2, n)
+  usr  <- graphics::par("usr")   # extremes of the plotting region (x1, x2, y1, y2)
+  pusr <- c(diff(usr[1:2]), diff(usr[3:4]))  # plot dimensions in user units (width, height)
+  asp <- (pusr[1] / pin[1]) / (pusr[2] / pin[2])  # y/x aspect ratio
 
-  if (is.lonlat) {
-    y <- (usr[3] + usr[4]) / 2
-    xaxp <- graphics::par("xaxp")
-    dx1 <- diff(xaxp[1:2]) / xaxp[3]
-    dm1 <- sp::spDistsN1(cbind(0, y), c(dx1, y), longlat=TRUE)
-    dm2 <- diff(pretty(c(0, dm1), 1)[1:2])
-    d <- (dx1 * dm2) / dm1
-    label <- paste(format(dm2), "km")
+  if (is.null(unit)) unit <- rep(as.character(NA), 2)
+  if (is.null(conv.fact)) conv.fact <- c(1, NA)
+  if (is.null(vert.exag)) {
+    vert.exag <- ifelse(asp > 20, TRUE, FALSE)
+  } else if (!is.logical(vert.exag)) {
+    asp <- vert.exag
+    vert.exag <- TRUE
+  }
+  if (is.null(offset)) offset <- 0
+  offset <- rep_len(offset, 2)
+
+  # calculate length of scale-bar axis in user units
+  dx <- diff(xaxp[1:2]) / xaxp[3]  # x distance between plot tick marks
+  if (longlat) {
+    y <- mean(usr[3:4])  # latitude at center of plot
+    ds1 <- sp::spDistsN1(cbind(0, y), c(dx, y), longlat=TRUE) # distance between plot tick marks in kilometers
+    ds1 <- ds1 * conv.fact[1]  # distance between plot tick marks in scale units
+    bp <- pretty(c(0, .Round(ds1)), min.n=2)  # pretty breakpoints
+    ds2 <- diff(range(bp))  # pretty distance between plot tick marks in scale units
+    lab_val <- ds2  # number label for last tick mark on scale
+    len <- dx * ds2 / ds1  # length of scale bar in user units
+    at <- len * bp / max(bp)  # scale tick-mark distances in user units
   } else {
-    d <- diff(pretty(usr[1:2]))[1]
-    label <- format(d, big.mark=",")
-    if (!is.null(unit)) label <- paste(label, unit)
+    at <- pretty(c(0, .Round(dx * conv.fact[1])), min.n=2)  # scale tick-mark locations in scale units
+    len <- diff(range(at))  # length of scale bar in user units
+    lab_val <- len  # number label for last tick mark on scale
+    at <- at / conv.fact[1]  # scale tick-mark distances in user units
+    len <- len / conv.fact[1]  # length of scale bar in user units
   }
 
-  padx <- 0.05 * (usr[2] - usr[1])
-  pady <- 0.05 * (usr[2] - usr[1]) / asp
+  # create scale-bar labels
+  lab_val <- format(lab_val, big.mark=",")
+  lab <-  ifelse(is.na(unit[1]), lab_val, paste(lab_val, unit[1]))
+
+  # determine user coordinate at origin of scale
+  sw <- graphics::strwidth("M", units="user", cex=1)  # string width in user units
+  sh <- graphics::strheight("M", units="user", cex=1)  # string height in user units
+  pad <- c(sw * ifelse(grepl("left$", loc), 2, 3),
+           sh * ifelse(grepl("^top",  loc), 3, 2))
   if (loc == "bottomleft") {
-    loc <- c(usr[1] + padx, usr[3] + pady)
+    xy <- c(usr[1] + pad[1], usr[3] + pad[2])
   } else if (loc == "topleft") {
-    loc <- c(usr[1] + padx, usr[4] - pady)
+    xy <- c(usr[1] + pad[1], usr[4] - pad[2])
   } else if (loc == "topright") {
-    loc <- c(usr[2] - padx - d, usr[4] - pady)
+    xy <- c(usr[2] - pad[1] - len, usr[4] - pad[2])
   } else if (loc == "bottomright") {
-    loc <- c(usr[2] - padx - d, usr[3] + pady)
+    xy <- c(usr[2] - pad[1] - len, usr[3] + pad[2])
   }
-  loc[1] <- loc[1] + offset[1] * (diff(usr[1:2]) / pin[1])
-  loc[2] <- loc[2] + offset[2] * (diff(usr[3:4]) / pin[2])
+  xy <- c(xy[1] + offset[1] * pusr[1] / pin[1], xy[2] + offset[2] * pusr[2] / pin[2])
 
-  divs <- 1L
-  for (i in 5:3) {
-    if (d %% i == 0) {
-      divs <- i
-      break
-    }
-  }
-  xat <- seq(loc[1], loc[1] + d, length.out=(divs + 1L))
-  tcl <- (diff(usr[1:2]) * 0.01) / asp
+  # draw axis and tick marks
+  xat <- xy[1] + at  # x of tick marks in user coordinates
+  tcl <- sh * 0.4  # y length of tick marks in user coordinates
+  graphics::lines(rbind(c(xy[1], xy[2]), c(xy[1] + len, xy[2])), lwd=0.5)
+  for (i in xat) graphics::lines(rbind(c(i, xy[2]), c(i, xy[2] + tcl)), lwd=0.5)
 
-  graphics::lines(rbind(c(loc[1], loc[2]), c(loc[1] + d, loc[2])), lwd=0.5)
-  for (i in xat) graphics::lines(rbind(c(i, loc[2]), c(i, loc[2] + tcl)), lwd=0.5)
-  graphics::text(loc[1], loc[2] + tcl, "0", adj=c(0.5, -0.5), cex=0.7)
-  graphics::text(loc[1] + d, loc[2] + tcl, label, adj=c(0.3, -0.5), cex=0.7)
+  # draw tick-mark labels at extremes of scale bar
+  lab_val_sw <- graphics::strwidth(lab_val, units="user", cex=0.7)
+  graphics::text(xy[1], xy[2] + tcl, "0", adj=c(0.5, -0.6), cex=0.7)
+  graphics::text(xy[1] + len - lab_val_sw / 2, xy[2] + tcl, lab, adj=c(0, -0.6), cex=0.7)
 
-  if (is.logical(lab.vert.exag)) {
-    add.label <- lab.vert.exag
-  } else {
-    add.label <- if (asp > 20) TRUE else FALSE
-  }
-  if (add.label) {
+  # draw label explaining vertical exaggeration beneath scale bar
+  if (vert.exag) {
     txt <- sprintf("VERTICAL EXAGGERATION x%s", asp)
-    graphics::text(loc[1] + d / 2, loc[2], txt, cex=0.7, pos=1)
+    graphics::text(xy[1] + len / 2, xy[2], txt, cex=0.7, pos=1)
   }
+}
+
+
+.Round <- function(x, base=c(1, 2, 5, 10)) {
+  checkmate::assertNumeric(x, finite=TRUE, min.len=1)
+  checkmate::assertIntegerish(base, lower=1, upper=10, any.missing=FALSE,
+                              min.len=2, unique=TRUE, sorted=TRUE)
+  n <- floor(log10(abs(x)))
+  m1 <- x / 10^n
+  vec <- c(0, utils::head(base, -1) + diff(base) / 2)
+  m2 <- base[findInterval(m1, vec)]
+  return(m2 * 10^n)
 }