+++ R/PrintFigure.R
@@ -15,9 +15,9 @@
 #'   Maximum number of rows and columns in the subfigure layout on a page in the output document.
 #' @param label 'character'.
 #'   String containing the LaTeX label anchor.
-#'   Subfigures are labeled using a concatenation of the \code{label} argument and a letter.
+#'   Subfigures are labeled using a concatenation of the \code{label} argument and an index number.
 #'   For example, specifying \code{label = "id"} for a figure composed of 3 subfigures results in:
-#'   \code{"id_a"}, \code{"id_b"}, and \code{"id_c"} labels.
+#'   labels \code{"id-1"}, \code{"id-2"}, and \code{"id-3"}.
 #' @param title 'character'.
 #'   String containing the figure caption.
 #' @param title_lof 'character'.
@@ -26,6 +26,18 @@
 #' @param headings 'character'.
 #'   Vector of subfigure captions, values are recycled as necessary
 #'   to match the vector length of the \code{fig} argument.
+#' @param pos 'character'.
+#'   String for the placement specifiers to be used in \code{\\begin{figure}[pos]}.
+#'   The specifiers can consist of the following characters in any order:
+#'   \itemize{
+#'     \item \code{"h"} place the float about at the same point it occurs in the source text;
+#'     \item \code{"t"} position at the top of the page;
+#'     \item \code{"b"} position at the bottom of the page;
+#'     \item \code{"p"} put on a special page for floats only; 
+#'     \item \code{"!"} override internal parameters LaTeX uses for determining float positions; and
+#'     \item \code{"H"} places the float at precisely the location in the source text.
+#'       Requires \code{\\usepackage{float}} in the LaTeX preamble.
+#'   }
 #'
 #' @details
 #'   Requires \code{\\usepackage{caption}} and \code{\\usepackage{subcaption}} in the LaTeX preamble.
@@ -53,8 +65,8 @@
 #'     "\\begin{document}\n",
 #'     "<<id, echo=FALSE, fig.width=3, fig.height=2, results='asis'>>=",
 #'     "par(mar=c(2.1, 2.1, 1.1, 1.1))",
-#'     "fig <- sprintf('plot(runif(%s))', 1:10)",
-#'     "headings <- sprintf('Subfigure caption, n=%s', 1:10)",
+#'     "fig <- sprintf('plot(runif(%s))', 1:9)",
+#'     "headings <- sprintf('Subfigure caption, n=%s', 1:9)",
 #'     "PrintFigure(fig, 3, 2, 'id', title='Figure caption', headings=headings)",
 #'     "@\n",
 #'     "\\end{document}",
@@ -67,7 +79,8 @@
 #' }
 #'
 
-PrintFigure <- function(fig, nr=1, nc=1, label="", title="", title_lof=title, headings="") {
+PrintFigure <- function(fig, nr=1, nc=1, label="", title="", title_lof=title, 
+                        headings="", pos="") {
 
   # check arguments
   checkmate::assertCharacter(fig, any.missing=FALSE, min.len=1)
@@ -77,6 +90,7 @@ PrintFigure <- function(fig, nr=1, nc=1, label="", title="", title_lof=title, he
   checkmate::assertString(title)
   checkmate::assertString(title_lof)
   checkmate::assertCharacter(headings, any.missing=FALSE, min.len=1, max.len=length(fig))
+  checkmate::assertString(pos)
 
   # total number of plots on all pages
   np <- length(fig)
@@ -85,13 +99,14 @@ PrintFigure <- function(fig, nr=1, nc=1, label="", title="", title_lof=title, he
   n <- nr * nc
   if (np < n) n <- np
 
-  # recycle subfigure captions
+  # recycle subfigure caption
   headings <- rep(headings, length.out=np)
+
+  # make subfigure labels unique
+  if (np > 1) label <- sprintf("%s-%s", label, 1:np)
 
   cat("\n")
   for (i in seq_len(np)) {
-    sublabel <- if (np > 1) sprintf("%s_%s", label, letters[i]) else label
-
     if (i == n) {
       caption <- strwrap(title, width=.Machine$integer.max)
     } else if (i > n && ((i %% n) == 0 || i == np)) {
@@ -103,7 +118,7 @@ PrintFigure <- function(fig, nr=1, nc=1, label="", title="", title_lof=title, he
     if (i == n + 1L) cat("\\captionsetup[figure]{list=no}\n\n")
 
     if ((i - 1L) %% n == 0) {
-      cat("\\begin{figure}\n")
+      cat(sprintf("\\begin{figure}[%s]\n", pos))
       if (i > 1) cat("  \\ContinuedFloat\n")
     } else if ((i - 1L) %% nc == 0) {
       cat("  \\par\\bigskip\n")
@@ -113,7 +128,7 @@ PrintFigure <- function(fig, nr=1, nc=1, label="", title="", title_lof=title, he
 
     if (np > 1) {
       cat(sprintf("  \\begin{subfigure}{%.2f\\textwidth}\n", 1 / nc))
-      cat(sprintf("    \\caption{{%s \\label{fig:%s}}}\n", headings[i], sublabel))
+      cat(sprintf("    \\caption{{%s \\label{fig:%s}}}\n", headings[i], label[i]))
     }
 
     # evaluate plotting commands
@@ -131,6 +146,7 @@ PrintFigure <- function(fig, nr=1, nc=1, label="", title="", title_lof=title, he
     }
 
     if (i %% n == 0 || i == np) cat("\\end{figure}\n\n")
+    
     if (i > n && i == np) cat("\\captionsetup[figure]{list=yes}\n\n")
   }